+++ src/pages/group/fee/FeeCreate.jsx
@@ -1,8 +1,7 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
 import FeeMemberSelect from './FeeMemberSelect';
-import Calendar from 'react-calendar';
-import 'react-calendar/dist/Calendar.css';
+import CustomCalendar from './fee_components/CustomCalendar';
 
 const Container = styled.div`
   width: 100%;
@@ -91,95 +90,6 @@ const Input = styled.input`
   }
 `;
 
-const CalendarContainer = styled.div`
-  top: 100%;
-  position: absolute;
-  right: 0;
-  z-index: 100;
-`;
-
-const StyledCalendar = styled(Calendar)`
-  background-color: #fff;
-  border: 1px solid ${({ theme }) => theme.colors.neutral[300]};
-  border-radius: 8px;
-  box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.1);
-
-  .react-calendar__navigation {
-    span {
-      font-size: 16px;
-      font-weight: 600;
-      color: ${({ theme }) => theme.colors.main[500]};
-    }
-  }
-
-  .react-calendar__navigation button:enabled:focus{
-    width: 15px;
-    background: ${({ theme }) => theme.colors.neutral[100]};
-  }
-
-  .react-calendar__tile {
-    padding: 8px;
-    abbr {
-      color: ${({ theme }) => theme.colors.neutral[500]};
-      text-decoration: none;
-    }
-  }
-
-  .react-calendar__tile--hasActive {
-    background: ${({ theme }) => theme.colors.main[500]};
-    border-radius: 12px;
-
-    abbr {
-      color: white;
-      text-decoration: none;
-    }
-  }
-
-  .react-calendar__tile--now {
-    background: ${({ theme }) => theme.colors.neutral[100]};
-    border-radius: 12px;
-
-    abbr {
-      color: ${({ theme }) => theme.colors.neutral[500]};
-      text-decoration: none;
-    }
-  }
-
-  .react-calendar__tile--active {
-    background: ${({ theme }) => theme.colors.main[500]};
-    border-radius: 12px;
-
-    abbr {
-      color: white;
-      text-decoration: none;
-    }
-  }
-
-  .react-calendar__tile:enabled:hover,
-  .react-calendar__tile:enabled:focus {
-    background: ${({ theme }) => theme.colors.main[100]};
-    border-radius: 12px;
-
-    abbr {
-      color: ${({ theme }) => theme.colors.neutral[500]};
-      text-decoration: none;
-    }
-  }
-
-  .react-calendar__month-view__weekdays {
-    abbr {
-      text-decoration: none;
-      color: ${({ theme }) => theme.colors.neutral[500]};
-    }
-  }
-
-  /* 저번 달 & 다음 달 일자 */
-  .react-calendar__month-view__days__day--neighboringMonth{
-    abbr {
-      color: ${({ theme }) => theme.colors.neutral[300]};
-    }
-  }
-`;
 
 const SelectMemberContainer = styled.div`
   width: 60%;
@@ -208,10 +118,6 @@ const FeeCreate = ({ closeFeeCreate }) => {
     if (showStartCalendar) setShowStartCalendar(false);
     if (showEndCalendar) setShowEndCalendar(false);
     if (showSelectFeeMember) setShowSelectFeeMember(false);
-  };
-
-  const handleCalendarClick = (e) => {
-    e.stopPropagation(); // 이벤트 버블링 중단
   };
 
   const handleFeeSelectClick = (e) => {
@@ -265,16 +171,13 @@ const FeeCreate = ({ closeFeeCreate }) => {
             }}
           />
           {showStartCalendar && (
-            <CalendarContainer onClick={handleCalendarClick}>
-              <StyledCalendar
-                onChange={(date) => {
-                  setStartDate(date);
-                  setShowStartCalendar(!showStartCalendar)
-                }}
-                value={startDate}
-                formatDay={(locale, date) => date.toLocaleString("en", {day: "numeric"})}
-              />
-            </CalendarContainer>
+            <CustomCalendar
+              value={startDate}
+              onChange={(date) => {
+                setStartDate(date);
+                setShowStartCalendar(false);
+              }}
+            />
           )}
         </FieldContainer>
 
@@ -290,16 +193,13 @@ const FeeCreate = ({ closeFeeCreate }) => {
             }}
           />
           {showEndCalendar && (
-            <CalendarContainer onClick={handleCalendarClick}>
-              <StyledCalendar
-                onChange={(date) => {
-                  setEndDate(date);
-                  setShowEndCalendar(!showEndCalendar)
-                }}
-                value={endDate}
-                formatDay={(locale, date) => date.toLocaleString("en", {day: "numeric"})}
-              />
-            </CalendarContainer>
+            <CustomCalendar
+              value={endDate}
+              onChange={(date) => {
+                setEndDate(date);
+                setShowEndCalendar(false);
+              }}
+            />
           )}
         </FieldContainer>
